Extracted developers route
