+++ routes/developers.js
@@ -1,10 +1,8 @@
 const express = require('express')
 const router = express.Router()
-const redis = require('redis') // redis connection
 const bodyParser = require('body-parser') // parses info from POST
 const methodOverride = require('method-override') // used to manipulate POST data
-
-const client = redis.createClient()
+const DeveloperHandler = require('../handlers/developers_handler.js')
 
 router.use(bodyParser.urlencoded({ extended: true }))
 router.use(methodOverride(function (req, res) {
@@ -17,63 +15,11 @@ router.use(methodOverride(function (req, res) {
 }))
 
 // Build the REST operations at the base of courses
-// This will be accessible from localhost:8000/developers
+// This will be accessible from localhost:3000/developers
 router.route('/:username')
-    .get(function (req, res) {
-      client.get(req.params.username, function (err, resp) {
-        if (err) {
-          setErrorResponse(res, err, 404)
-        } else if (resp === null) {
-          setErrorResponse(res, 'The username provided does not match any username. Please make sure that you are signed up as an authorized ciena developer on www.cienadevelopers.com', 500)
-        } else {
-          setStandardResponse(res, resp)
-        }
-      })
-    })
-    .post(function (req, res) {
-      client.set(req.params.username, req.body.token, function (err, resp) {
-        if (err) {
-          setErrorResponse(res, err, 404)
-        } else {
-          setStandardResponse(res, resp)
-        }
-      })
-    })
-    .put(function (req, res) {
-      client.set(req.params.username, req.body.token, function (err, resp) {
-        if (err) {
-          setErrorResponse(res, err, 404)
-        } else {
-          setStandardResponse(res, resp)
-        }
-      })
-    })
-    .delete(function (req, res) {
-      client.del(req.params.username, function (err, resp) {
-        if (err) {
-          setErrorResponse(res, err, 404)
-        } else {
-          setStandardResponse(res, resp)
-        }
-      })
-    })
-
-function setErrorResponse (res, err, statusCode) {
-  res.status(statusCode)
-  res.format({
-    json: function () {
-      res.json({error: err})
-    }
-  })
-}
-
-function setStandardResponse (res, resp) {
-  res.status(200)
-  res.format({
-    json: function () {
-      res.json(resp)
-    }
-  })
-}
+    .get(DeveloperHandler.get)
+    .post(DeveloperHandler.post)
+    .put(DeveloperHandler.put)
+    .delete(DeveloperHandler.delete)
 
 module.exports = router
